feat: Refactor file upload functionality
