+++ backend/schemas/File.py
@@ -1,0 +1,4 @@
+class FileUploadError(Exception):
+    """Custom exception for file upload errors"""
+    pass
+
